Fixes #20 - Fix 1.12 deprecation
